add NonObjectTypeTrait
