+++ src/Type/FalseBooleanType.php
@@ -2,63 +2,17 @@
 
 namespace PHPStan\Type;
 
-use PHPStan\Analyser\Scope;
-use PHPStan\Reflection\ClassConstantReflection;
-use PHPStan\Reflection\MethodReflection;
-use PHPStan\Reflection\PropertyReflection;
 use PHPStan\TrinaryLogic;
+use PHPStan\Type\Traits\NonObjectTypeTrait;
 
 class FalseBooleanType implements BooleanType
 {
 
+	use NonObjectTypeTrait;
+
 	public function describe(): string
 	{
 		return 'false';
-	}
-
-	public function canAccessProperties(): bool
-	{
-		return false;
-	}
-
-	public function hasProperty(string $propertyName): bool
-	{
-		return false;
-	}
-
-	public function getProperty(string $propertyName, Scope $scope): PropertyReflection
-	{
-		throw new \PHPStan\ShouldNotHappenException();
-	}
-
-	public function canCallMethods(): bool
-	{
-		return false;
-	}
-
-	public function hasMethod(string $methodName): bool
-	{
-		return false;
-	}
-
-	public function getMethod(string $methodName, Scope $scope): MethodReflection
-	{
-		throw new \PHPStan\ShouldNotHappenException();
-	}
-
-	public function canAccessConstants(): bool
-	{
-		return false;
-	}
-
-	public function hasConstant(string $constantName): bool
-	{
-		return false;
-	}
-
-	public function getConstant(string $constantName): ClassConstantReflection
-	{
-		throw new \PHPStan\ShouldNotHappenException();
 	}
 
 	/**
@@ -122,11 +76,6 @@ class FalseBooleanType implements BooleanType
 	public function isCallable(): TrinaryLogic
 	{
 		return TrinaryLogic::createNo();
-	}
-
-	public function isClonable(): bool
-	{
-		return false;
 	}
 
 	public static function __set_state(array $properties): Type
